+++ tests/testthat/test_Apply.R
@@ -14,6 +14,62 @@ library(tidyr)
 
 
 ###########################################################
+context("Element apply")
+###########################################################
+
+test_that("elementapply_byname works as expected", {
+  half <- function(x){
+    x/2
+  }
+  m <- matrix(c(1:4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+    setrowtype("row") %>% setcoltype("col")
+  expect_equal(elementapply_byname(FUN = half, a = m, row = "r1", col = "c1"), 
+               matrix(c(0.5, 2, 3, 4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col"))
+  expect_equal(elementapply_byname(FUN = half, a = m, row = "r2", col = "c1"), 
+               matrix(c(1, 1, 3, 4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col"))
+  expect_equal(elementapply_byname(FUN = half, a = m, row = "r1", col = "c2"), 
+               matrix(c(1, 2, 1.5, 4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col"))
+  expect_equal(elementapply_byname(FUN = half, a = m, row = "r2", col = "c2"), 
+               matrix(c(1, 2, 3, 2), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col"))
+  
+  # Try some pathological cases.
+  expect_null(elementapply_byname(half, a = NULL, row = "r", col = "c"))
+  expect_error(elementapply_byname(half, a = m, row = "bogus", col = "c1"), 
+               "subscript out of bounds")
+  
+  # Try indices for rows and columns.
+  expect_equal(elementapply_byname(half, a = m, row = 1, col = 1), 
+               matrix(c(0.5, 2, 3, 4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col"))
+  # Try a mix of names and indices
+  expect_equal(elementapply_byname(half, a = m, row = 1, col = "c2"), 
+               matrix(c(1, 2, 1.5, 4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col"))
+  # Try with .FUNdots
+  divide <- function(x, divisor){
+    x/divisor
+  }
+  expect_equal(elementapply_byname(divide, a = m, row = 1, col = 1, .FUNdots = list(divisor = 2)), 
+               matrix(c(0.5, 2, 3, 4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col"))
+  expect_equal(elementapply_byname(divide, a = m, row = 1, col = 1, .FUNdots = list(divisor = 10)), 
+               matrix(c(0.1, 2, 3, 4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col"))
+  
+  # Should also work with lists
+  l <- list(m, m)
+  expected <- matrix(c(10, 2, 3, 4), nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
+                 setrowtype("row") %>% setcoltype("col")
+  expect_equal(elementapply_byname(divide, a = l, row = 1, col = 1, .FUNdots = list(divisor = 0.1)), 
+               list(expected, expected))
+})
+
+
+###########################################################
 context("Unary apply")
 ###########################################################