+++ NEWS.md
@@ -1,3 +1,9 @@
+* New function `elementapply_byname()` applies a function to an element
+  of a matrix specified by `row` and `col` arguments.
+* Breaking change: `elementproduct_byname()` changed to `hadamardproduct_byname()`
+  to avoid name collision with `elementapply_byname()`.
+
+
 # matsbyname 0.4.4 (2019-01-02)
 
 * Added tests to achieve 100% code coverage.